Moved images to documentation folder, code to src folder
